test(dashboard): exercise server routes directly
